Get the date from the last commit.
As we are in an early development phase, the generated documentation is marked
with the current git tag and the date when the documentation was generated.

To make the last documentation available publicly as soon as possible, a bot
will be configured to generate the documentation daily. In order to prevent too
many commits, we'll base the date on the last commit instead :
it means that generated documentations will be committed and pushed only if the
original documentations *and* the converter have changed.