Fix some Example parsing issue, leading into missing text
llange on github reported that part of text was missing for some examples.
This is is due to the regex used to detect such keywords, which only accepted
"Example" at the beginning of a line. Any text before it was simply ignored.

This patch considers that the use case is uncommon and disable the "Example"
rendering in this case, which fixes issue #9.