Early preview of the new documentation files
HAProxy has split its documentation in several ones but I was late to adapt
the converter. It's now time to fix this issue.
Issues will arise at first, the goal is to identify them quickly and fix them.

It also implies a new convention for files and directories naming. Now, each
version will be stored in its own directory.