Add a menu to navigate between documentation files
Now we have a list of files, it's possible to render a menu to access to each
ones. The legacy menu has been removed (most of the informations are available
on the HAProxy home page).

Maybe it will require to reintroduce a menu to easily switch from one version
to another.