Remove the search engine
This was a funny tool but the library used for the purpose looks to be
unmaintained. Some users also reported it was counterintuitive for some
use cases. For example, only known keywords were found but users expected
a full text search in the whole document.

For now, we can remove the feature and consider that the full text search
provided by the browser is enough.