keyword detection: consistently allow spaces in argument names
When keywords are parsed, spaces in argument names are not handled
consistently.

The following is parsed correctly:
  group <group name>

While the following isn't:
  unset-var(<var name>)

With this commit the latter would be parsed. Note that it would also
allow things such as:
  ipmask(<mask4>, [<mask6>])
  hmac(<algorithm>, <key>)

Both of these are not correct in that a space after the period
would prevent the config from being parsed by haproxy, but these
examples were in the actual documentation not long ago, and when
parsed by haproxy-dconv this makes them not searchable and not
correctly formatted in the html.